For speculative execution, writers cannot use the final output filenames. They must use temporary files instead and only the first successful jobs will move its files to the final location. However, there is a bug in the renaming code (see Treex/Core/t/writers.t) so let's disable it. TODO: fix it properly.
I tried a proper fix without success, I give up / postpone it.


git-svn-id: https://svn.ms.mff.cuni.cz/svn/tectomt_devel/trunk/treex@9389 7cf035df-8472-40d0-8fbe-a733a24d2855

+++ lib/Treex/Block/Read/ProducerReader.pm
@@ -107,15 +107,14 @@ has _finished_jobs => (
 
 has log_file => (
     isa => 'Str',
-    is => 'ro',
+    is  => 'ro',
 );
 
 has 'survive' => (
-    is            => 'rw',
-    isa           => 'Bool',
-    default       => 0,
+    is      => 'rw',
+    isa     => 'Bool',
+    default => 0,
 );
-
 
 sub BUILD {
     my $self = shift;
@@ -124,12 +123,13 @@ sub BUILD {
     my $result    = 1;
     my $reconnect = 0;
 
-    open(my $fh_log, ">:encoding(UTF-8)", $self->log_file) or croak $! . " - " . $self->log_file;
+    open( my $fh_log, ">:encoding(UTF-8)", $self->log_file ) or croak $! . " - " . $self->log_file;
 
     do {
         POE::Session->create(
             inline_states => {
                 _start => sub {
+
                     # Start the server.
                     $reconnect = 0;
                     $_[HEAP]{server} = POE::Wheel::SocketFactory->new(
@@ -155,12 +155,12 @@ sub BUILD {
 
                     # Shut down server.
                     my ( $operation, $errnum, $errstr ) =
-                      @_[ ARG0, ARG1, ARG2 ];
+                        @_[ ARG0, ARG1, ARG2 ];
                     log_warn "MSG: Server $operation error $errnum: $errstr\n";
 
                     # address already used
                     if ( $errnum == 98 ) {
-                        $reconnect = 1;
+                        $reconnect              = 1;
                         $Treex::Core::Run::PORT = int( 30000 + rand(32000) );
                     }
                     delete $_[HEAP]{server};
@@ -169,36 +169,40 @@ sub BUILD {
 
                     # Handle client input.
                     my ( $kernel, $sender, $heap, $input, $wheel_id ) =
-                      @_[ KERNEL, SESSION, HEAP, ARG0, ARG1 ];
+                        @_[ KERNEL, SESSION, HEAP, ARG0, ARG1 ];
                     my ( $jobid, $function ) = split( /\t/, $$input );
 
-                    my $msg = "Client $jobid: $function";
+                    my $msg          = "Client $jobid: $function";
                     my $finished_str = "__finished__";
                     $result = $finished_str;
-                    
-                    if ( ($self->survive == 0 && $self->_crashed_jobs == 0) ||
-                         ($self->survive == 1 && $self->_crashed_jobs < $self->jobs )
-                    ) {
+
+                    if (( $self->survive == 0 && $self->_crashed_jobs == 0 )
+                        ||
+                        ( $self->survive == 1 && $self->_crashed_jobs < $self->jobs )
+                        )
+                    {
                         if ( $function =~ /cmd_(.*)(\t(.*))?/ ) {
                             my $act_cmd = $1;
                             if ( $act_cmd eq "started" ) {
                                 $result = "started";
-                            } elsif ( $act_cmd eq "fatalerror" ) {
+                            }
+                            elsif ( $act_cmd eq "fatalerror" ) {
                                 $result = "fatalerror";
 
                                 # zjisti, ktery to je
                                 my $finished_file = $self->_jobs->{$jobid};
-    
+
                                 if ($finished_file) {
-    
+
                                     # error during document processing
-    
+
                                     $msg .= "; Crashed $finished_file";
                                     if ( $self->_files->{$finished_file}->{'finished'} ) {
                                         $msg .= "; Already finished!!!";
-                                        $self->_process_created_files($jobid, 0, 1);
-                                    } else {
-                                        $self->_process_created_files($jobid, 0, 1);
+                                        $self->_process_created_files( $jobid, 0, 1 );
+                                    }
+                                    else {
+                                        $self->_process_created_files( $jobid, 0, 1 );
                                         $self->_files->{$finished_file}->{'crashed'}++;
                                         $msg .= "; Crashed " . $self->_files->{$finished_file}->{'crashed'} . " times";
                                         if ( $self->_files->{$finished_file}->{'crashed'} > $self->_submitted_limit - 1 ) {
@@ -206,112 +210,117 @@ sub BUILD {
                                             $self->_set_finished_count( $self->_finished_count + 1 );
                                         }
                                     }
-                                } else {
+                                }
+                                else {
                                     $msg .= "; Crashed during loading";
                                 }
                                 $self->status->{"info_fatal_job"} = $jobid;
                                 $self->status->{"info_fatal_doc"} = $finished_file;
-    
+
                                 # increase number of crashed jobs
                                 $self->status->{'info_fatalerror'} = $self->status->{'info_fatalerror'} + 1;
                                 $self->_set_crashed_jobs( $self->_crashed_jobs + 1 );
                                 my $remaining_jobs = $self->jobs - $self->_finished_jobs - $self->_crashed_jobs;
+
                                 #TODO: hack
                                 if ( $remaining_jobs < 0 ) {
                                     $remaining_jobs = 0;
                                 }
                                 log_info( "Remains " . $remaining_jobs . " jobs (" . $self->_crashed_jobs . " crashed) out of " . $self->jobs );
-    
+
                                 $self->status->{'info_crashed_jobs'} = $self->_crashed_jobs;
                                 if ( $remaining_jobs == 0 ) {
                                     $self->_mark_as_finished();
                                 }
                             }
-    
+
                             # delete auxiliary created files
                             if ( $act_cmd eq "finished" ) {
-    
+
                                 my $orig_dir =
-                                  $self->workdir . '/output__JOB__' . $jobid . '/';
-    
+                                    $self->workdir . '/output__JOB__' . $jobid . '/';
+
                                 #qx(rm -rf $orig_dir);
-    
+
                                 for my $writer ( @{ $self->writers } ) {
                                     my $path = $writer->path;
                                     if ($path) {
                                         $path =~ s/\/+$//;
                                         $path .= "__JOB__" . $jobid;
-    
+
                                         #log_info("Client $jobid: DELETING $path");
                                         #qx(rm -rf $path);
                                     }
                                 }
-    
+
                                 $self->_set_finished_jobs( $self->_finished_jobs + 1 );
                             }
-    
+
                             $self->status->{ "job_" . $jobid . "_" . $act_cmd } = 1;
                         }
                         else {
                             my $received = $self->reader->$function();
-    
+
                             # mark job as finished
                             my $finished_file = $self->_jobs->{$jobid};
-    
+
                             if ($finished_file) {
                                 $msg .= "; Finished $finished_file";
                                 if ( $self->_files->{$finished_file}->{'finished'} ) {
                                     $msg .= "; Already finished!!!";
-                                    $self->_process_created_files($jobid, 1, 0);
-                                } else {
-                                    $self->_process_created_files($jobid, 0, 0);
-    
+                                    $self->_process_created_files( $jobid, 1, 0 );
+                                }
+                                else {
+                                    $self->_process_created_files( $jobid, 0, 0 );
+
                                     $self->_files->{$finished_file}->{'finished'} = time();
-    
+
                                     $self->status->{ "doc_" . $finished_file . "_finished" } = $jobid;
-    
+
                                     $self->_set_finished_count( $self->_finished_count + 1 );
                                 }
                             }
-    
+
                             if ($received) {
                                 $result = { 'result' => $received };
                                 $self->_set_doc_number( $self->doc_number + 1 );
-    
+
                                 if ( $self->reader->isa('Treex::Block::Read::BaseReader') ) {
                                     $result->{file_number} = $self->reader->file_number;
-                                } elsif ( $self->reader->isa('Treex::Block::Read::BaseAlignedReader') ) {
-                                    $result->{doc_number} = $self->doc_number;
+                                }
+                                elsif ( $self->reader->isa('Treex::Block::Read::BaseAlignedReader') ) {
+                                    $result->{doc_number}  = $self->doc_number;
                                     $result->{file_number} = $self->reader->_file_number;
                                 }
-    
+
                                 # mark job as started
                                 $self->status->{ "doc_" . $self->doc_number . "_started" } = $jobid;
-    
+
                                 $self->_files->{ $result->{file_number} } = {
                                     'started'   => time(),
                                     'job'       => $jobid,
                                     'result'    => $result,
                                     'submitted' => 1
                                 };
-    
+
                                 $msg .= "; Assigned: " . $result->{file_number};
-    
+
                                 $self->_set_file_count( $result->{file_number} );
                             }
                             else {
                                 $self->_set_total_file_count( $self->_file_count );
-    
+
                                 my $not_finished = 0;
-    
-                              SUBMITS:
+
+                                SUBMITS:
                                 for ( my $submitted = $self->_max_submitted; $submitted < $self->_submitted_limit; $submitted++ ) {
-    
+
                                     #$not_finished = 0;
                                     for ( my $i = $self->_file_count; $i > 1; $i-- ) {
                                         if ( !$self->_files->{$i}->{finished} ) {
+
                                             #        $not_finished++;
-                                            if ( $self->_files->{$i}->{submitted} < $submitted )  {
+                                            if ( $self->_files->{$i}->{submitted} < $submitted ) {
                                                 $self->_files->{$i}->{submitted} += 1;
                                                 $result = $self->_files->{$i}->{result};
                                                 $msg .= "; Assigned: " . $result->{file_number} . "; AGAIN: " . $self->_files->{$i}->{submitted};
@@ -321,7 +330,7 @@ sub BUILD {
                                         $self->_set_max_submitted($submitted);
                                     }
                                 }
-    
+
                                 if ( $self->_finished_count == $self->_file_count )
                                 {
                                     $self->_mark_as_finished();
@@ -329,7 +338,7 @@ sub BUILD {
                                     $_[KERNEL]->stop();
                                 }
                             }
-    
+
                             # remember wich job is working on which file
                             if ( ref($result) ) {
                                 $self->_jobs->{$jobid} = $result->{file_number};
@@ -337,6 +346,7 @@ sub BUILD {
                         }
                     }
                     $_[HEAP]{client}{$wheel_id}->put( \$result );
+
                     #log_info($msg);
                     print $fh_log running_time() . "\t" . $msg . "\n";
 
@@ -347,7 +357,7 @@ sub BUILD {
                     my $wheel_id = $_[ARG3];
                     delete $_[HEAP]{client}{$wheel_id};
                 },
-            }
+                }
         );
         POE::Kernel->run();
     } while ( $reconnect == 1 );
@@ -369,8 +379,9 @@ sub _mark_as_finished {
     }
 
     for my $id ( 1 .. $max_file ) {
-        if (   !defined( $self->status->{ "doc_" . $id . "_finished" } )
-            && !defined( $self->status->{ "doc_" . $id . "_fatalerror" } ) )
+        if (!defined( $self->status->{ "doc_" . $id . "_finished" } )
+            && !defined( $self->status->{ "doc_" . $id . "_fatalerror" } )
+            )
         {
             $self->status->{ "doc_" . $id . "_skipped" } = 1;
         }
@@ -401,71 +412,74 @@ sub _process_created_files {
     my ( $self, $jobid, $delete, $error ) = @_;
 
     my $finished_file = $self->_jobs->{$jobid};
-    
+
     my $output_dir = "output";
-    if ( $error ) {
+    if ($error) {
         $output_dir = "error";
     }
 
-    my $global_orig_dir   = Treex::Core::Run::construct_output_dir_name(
-        $self->workdir . '/output', $jobid, $self->host, $self->port);
+    my $global_orig_dir = Treex::Core::Run::construct_output_dir_name(
+        $self->workdir . '/output', $jobid, $self->host, $self->port
+    );
     my $global_target_dir = $self->workdir . "/$output_dir/";
 
     my @cmds = ();
 
-    for my $writer ( @{ $self->writers } ) {
-        my $path = $writer->path;
+    if ($Treex::Core::Run::SPECULATIVE_EXECUTION) {
+        for my $writer ( @{ $self->writers } ) {
+            my $path = $writer->path;
 
-        #log_warn("$writer - PATH: " . $writer->path . "; TO: " . $writer->to);
-        if ( !$path ) {
-            $path = "./";
-        }
-
-        my $target_path = $path;
-        $target_path .= "/";
-
-        if ( $path ne "./" ) {
-            $path =~ s/\/+$//;
-        }
-
-        $path = Treex::Core::Run::construct_output_dir_name($path, $jobid, $self->host, $self->port);
-
-        #log_warn("BEFORE: " . join(", ", glob $path . "/*"));
-
-        if ( $writer->to && $writer->to eq "-" ) {
-
-            my $out_file = $global_target_dir . '/'
-              . sprintf( "doc%07d.stdout", $finished_file );
-            my $in_file = $path . "/*";
-            my $try     = 0;
-
-            if ( $delete ) {
-                push( @cmds, "rm $in_file 2>/dev/null" );
-            } else {
-                push( @cmds, "mv $in_file $out_file  2>/dev/null" );
+            #log_warn("$writer - PATH: " . $writer->path . "; TO: " . $writer->to);
+            if ( !$path ) {
+                $path = "./";
             }
 
-        }
-        else {
+            my $target_path = $path;
+            $target_path .= "/";
 
-            #log_info("Client $jobid: mv $path/* $target_path/");
-            if ( $delete ) {
-                qx(rm $path/* 2>/dev/null);
-            } else {
-                qx(mv $path/* $target_path/ 2>/dev/null);
+            if ( $path ne "./" ) {
+                $path =~ s/\/+$//;
+            }
+
+            $path = Treex::Core::Run::construct_output_dir_name( $path, $jobid, $self->host, $self->port );
+
+            #log_warn("BEFORE: " . join(", ", glob $path . "/*"));
+
+            if ( $writer->to && $writer->to eq "-" ) {
+
+                my $out_file = $global_target_dir . '/'
+                    . sprintf( "doc%07d.stdout", $finished_file );
+                my $in_file = $path . "/*";
+                my $try     = 0;
+
+                if ($delete) {
+                    push( @cmds, "rm $in_file 2>/dev/null" );
+                }
+                else {
+                    push( @cmds, "mv $in_file $out_file  2>/dev/null" );
+                }
+
+            }
+            else {
+
+                #log_info("Client $jobid: mv $path/* $target_path/");
+                if ($delete) {
+                    qx(rm $path/* 2>/dev/null);
+                }
+                else {
+                    qx(mv $path/* $target_path/ 2>/dev/null);
+                }
             }
         }
-
-        #log_warn("AFTER: " . join(", ", glob $path . "/*"));
-
     }
 
     # move all related files to output folder
     #log_warn("GLOBAL BEFORE: " . join(", ", glob $global_orig_dir . "/*"));
     #log_warn("mv $global_orig_dir/* $global_target_dir");
-    if ( $delete ) {
+    if ($delete) {
         qx(rm -f $global_orig_dir/* 2>/dev/null);
-    } else {
+    }
+    else {
         qx(mv $global_orig_dir/* $global_target_dir 2>/dev/null);
     }
